# docs/openapi.yml
openapi: 3.0.0
info:
  title: Blog API
  description: API for a blogging platform.
  version: 1.0.0
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: http://localhost:8080/v1
    description: Development server
paths:
...
  /images/upload:
    post:
      summary: Upload an image
      description: |
        ユーザーがアップロードした画像を受け取り、サーバーに保存し、画像のURLを返します。
        - **ファイルストレージ**: 画像はサーバーのファイルシステムに保存されます。
        - **セキュリティ**: 画像のファイルタイプとサイズを検証します。
        - **認証**: このエンドポイントは認証ユーザーのみがアクセスできます。
      operationId: uploadImage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ImageUploadRequest'
      responses:
        '200':
          description: 画像が正常にアップロードされた場合のレスポンス
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageUploadResponse'
        '400':
          description: 不正なリクエスト
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 認証が必要な場合
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    ImageUploadRequest:
      type: object
      required:
        - image
      properties:
        image:
          type: string
          format: binary
          description: アップロードする画像ファイル。
      description: 画像アップロード用のマルチパートフォームデータ。
    ImageUploadResponse:
      type: object
      properties:
        url:
          type: string
          format: uri
          description: アップロードされた画像のURL。
      required:
        - url
      description: 画像アップロード成功時のレスポンス。
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: エラーメッセージ。
        code:
          type: integer
          format: int32
          description: エラーコード。
      required:
        - message
        - code
      description: エラーレスポンスの形式。
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT